test(metered): trim redundant metering coverage
