+++ crates/vm/src/arch/execution_mode/metered/mod.rs
@@ -1,3 +1,4 @@
 pub mod ctx;
 pub mod memory_ctx;
+mod page_tracker;
 pub mod segment_ctx;
